[LC-1958] Move the WASM URL guard out of the nx test graph
Asserting this inside didkit-plugin:test bailed every task depending on
didkit-plugin: skipped test suites went from 10 to 30, which suppressed
credential-library, learn-card-base, learn-cloud-*, lca-api-* and even the
ler-rs regression added alongside it. A stale release artifact should not cost
the review its test signal.

Run it as a standalone pure-Node lint job instead, mirroring the existing
check-i18n-keys scan. Same actionable message and the same hard block, but the
test graph runs to completion.
